fix smallint

+++ er/avaliacao.sql
@@ -5,7 +5,7 @@ CREATE SCHEMA avaliacao;
 CREATE TABLE avaliacao.questionario(
   id SERIAL NOT NULL PRIMARY KEY,
   nome VARCHAR(255) NOT NULL,
-  subfase_id SMALLINT NOT NULL REFERENCES macrocontrole.etapa (id),
+  subfase_id INTEGER NOT NULL REFERENCES macrocontrole.etapa (id),
   UNIQUE (subfase_id)
 );
 
@@ -13,27 +13,27 @@ CREATE TABLE avaliacao.pergunta(
   id SERIAL NOT NULL PRIMARY KEY,
   texto VARCHAR(255) NOT NULL,
   ordem INTEGER NOT NULL, -- as perguntas são ordenadas dentro de um questionário
-  questionario_id SMALLINT NOT NULL REFERENCES avaliacao.questionario (id)
+  questionario_id INTEGER NOT NULL REFERENCES avaliacao.questionario (id)
 );
 
 CREATE TABLE avaliacao.opcao(
   id SERIAL NOT NULL PRIMARY KEY,
   texto VARCHAR(255) NOT NULL,
   ordem INTEGER NOT NULL, -- as opções são ordenadas dentro de uma pergunta
-  pergunta_id SMALLINT NOT NULL REFERENCES avaliacao.pergunta (id)
+  pergunta_id INTEGER NOT NULL REFERENCES avaliacao.pergunta (id)
 );
 
 CREATE TABLE avaliacao.resposta_questionario(
   id SERIAL NOT NULL PRIMARY KEY,
   data timestamp with time zone NOT NULL,
-  atividade_id SMALLINT NOT NULL REFERENCES macrocontrole.atividade (id)
+  atividade_id INTEGER NOT NULL REFERENCES macrocontrole.atividade (id)
 );
 
 CREATE TABLE avaliacao.resposta(
   id SERIAL NOT NULL PRIMARY KEY,
-  pergunta_id SMALLINT NOT NULL REFERENCES avaliacao.pergunta (id),
-  opcao_id SMALLINT NOT NULL REFERENCES avaliacao.opcao (id),
-  resposta_questionario_id SMALLINT NOT NULL REFERENCES avaliacao.resposta_questionario (id)
+  pergunta_id INTEGER NOT NULL REFERENCES avaliacao.pergunta (id),
+  opcao_id INTEGER NOT NULL REFERENCES avaliacao.opcao (id),
+  resposta_questionario_id INTEGER NOT NULL REFERENCES avaliacao.resposta_questionario (id)
 );
 
 COMMIT;
